add smoothing
motion is pretty natural, but when you aren't moving, fluctuations in the models detection of the face cause some jarring movements. Since these seems to be pretty randomly distributed, taking your change in position to be the average of the last few positions cancels out much of the noise. It would be nice if it could then somehow be reduced to 0 if it appears that you're not moving, but I haven't gotten that to work yet exactly. Modifying smoothingConstant to 1 replicates the behavior from before this commit.